+++ sdk/eventhub/azure-eventhub/azure/eventhub/_consumer_client.py
@@ -10,6 +10,7 @@ from ._client_base import ClientBase
 from ._consumer import EventHubConsumer
 from ._constants import ALL_PARTITIONS
 from ._eventprocessor.event_processor import EventProcessor
+from ._eventprocessor.common import LoadBalancingStrategy
 
 
 if TYPE_CHECKING:
@@ -84,11 +85,22 @@ class EventHubConsumerClient(ClientBase):
      If a checkpoint store is not provided, the checkpoint will be maintained internally
      in memory, and the `EventHubConsumerClient` instance will receive events without load-balancing.
     :paramtype checkpoint_store: ~azure.eventhub.CheckpointStore
-    :keyword load_balancing_strategy: TODO: add description
-    :paramtype load_balancing_strategy: ~azure.eventhub.LoadBalancingStrategy
     :keyword float load_balancing_interval: When load-balancing kicks in. This is the interval, in seconds,
      between two load-balancing evaluations. Default is 10 seconds.
-    :keyword float partition_ownership_expiration_interval: TODO: add description
+    :keyword float partition_ownership_expiration_interval: A partition ownership will expire after this number
+     of seconds. Every load-balancing evaluation will automatically extend the ownership expiration time.
+     Default is 6 * load_balancing_interval.
+    :keyword str or ~azure.eventhub.LoadBalancingStrategy load_balancing_strategy: When load-balancing kicks in,
+     it will use this strategy to claim and balance the partition ownership.
+     Use "greedy" or `LoadBalancingStrategy.GREEDY` for the greedy strategy, which, for every
+     load-balancing evaluation, will grab all partitions that are not claimed by any other `EventHubConsumerClient`.
+     Use "balanced" or `LoadBalancingStrategy.BALANCED` for the balanced strategy, which, for every load-balancing
+     evaluation, claims only one partition that is not claimed by other `EventHubConsumerClient`.
+     If all partitions of an EventHub are claimed by other `EventHubConsumerClient` and this client has claimed
+     too fewer partitions, this client will steal one partition from other clients for every load-balancing
+     evaluation regardless of the load balancing strategy.
+     Greedy strategy is used by default.
+
 
     .. admonition:: Example:
 
@@ -110,7 +122,13 @@ class EventHubConsumerClient(ClientBase):
     ):
         # type: (...) -> None
         self._checkpoint_store = kwargs.pop("checkpoint_store", None)
-        self._load_balancing_interval = kwargs.pop("load_balancing_interval", 10)
+        self._load_balancing_interval = kwargs.pop("load_balancing_interval", None) or 10
+        self._partition_ownership_expiration_interval = kwargs.pop(
+            "partition_ownership_expiration_interval", None
+        ) or 6*self._load_balancing_interval
+        load_balancing_strategy = kwargs.pop("load_balancing_strategy", None)
+        self._load_balancing_strategy = LoadBalancingStrategy(load_balancing_strategy) if load_balancing_strategy \
+            else LoadBalancingStrategy.GREEDY
         self._consumer_group = consumer_group
         network_tracing = kwargs.pop("logging_enable", False)
         super(EventHubConsumerClient, self).__init__(
@@ -196,11 +214,22 @@ class EventHubConsumerClient(ClientBase):
          If a checkpoint store is not provided, the checkpoint will be maintained internally
          in memory, and the `EventHubConsumerClient` instance will receive events without load-balancing.
         :paramtype checkpoint_store: ~azure.eventhub.CheckpointStore
-        :keyword load_balancing_strategy: TODO: add description
-        :paramtype load_balancing_strategy: ~azure.eventhub.LoadBalancingStrategy
         :keyword float load_balancing_interval: When load-balancing kicks in. This is the interval, in seconds,
          between two load-balancing evaluations. Default is 10 seconds.
-        :keyword float partition_ownership_expiration_interval: TODO: add description
+        :keyword float partition_ownership_expiration_interval: A partition ownership will expire after this number
+         of seconds. Every load-balancing evaluation will automatically extend the ownership expiration time.
+         Default is 6 * load_balancing_interval.
+        :keyword str or ~azure.eventhub.LoadBalancingStrategy load_balancing_strategy: When load-balancing kicks in,
+         it will use this strategy to claim and balance the partition ownership.
+         Use "greedy" or `LoadBalancingStrategy.GREEDY` for the greedy strategy, which, for every
+         load-balancing evaluation, will grab all partitions that are not claimed by any other `EventHubConsumerClient`.
+         Use "balanced" or `LoadBalancingStrategy.BALANCED` for the balanced strategy, which, for every load-balancing
+         evaluation, claims only one partition that is not claimed by other `EventHubConsumerClient`.
+         If all partitions of an EventHub are claimed by other `EventHubConsumerClient` and this client has claimed
+         too fewer partitions, this client will steal one partition from other clients for every load-balancing
+         evaluation regardless of the load balancing strategy.
+         Greedy strategy is used by default.
+
         :rtype: ~azure.eventhub.EventHubConsumerClient
 
         .. admonition:: Example:
@@ -257,6 +286,8 @@ class EventHubConsumerClient(ClientBase):
                 on_event,
                 checkpoint_store=self._checkpoint_store,
                 load_balancing_interval=self._load_balancing_interval,
+                load_balancing_strategy=self._load_balancing_strategy,
+                partition_ownership_expiration_interval=self._partition_ownership_expiration_interval,
                 initial_event_position=initial_event_position,
                 initial_event_position_inclusive=initial_event_position_inclusive,
                 **kwargs