+++ sdk/eventhub/azure-eventhub/azure/eventhub/aio/_eventprocessor/_ownership_manager.py
@@ -8,6 +8,8 @@ import random
 from collections import Counter, defaultdict
 from typing import List, Iterable, Optional, Dict, Any, Union, TYPE_CHECKING
 
+from ..._eventprocessor.common import LoadBalancingStrategy
+
 if TYPE_CHECKING:
     from .checkpoint_store import CheckpointStore
     from .._consumer_client_async import EventHubConsumerClient
@@ -32,6 +34,7 @@ class OwnershipManager(object):  # pylint:disable=too-many-instance-attributes
         owner_id: str,
         checkpoint_store: Optional["CheckpointStore"],
         ownership_timeout: float,
+        load_balancing_strategy,  # type: LoadBalancingStrategy
         partition_id: Optional[str],
     ):
         self.cached_parition_ids = []  # type: List[str]
@@ -45,8 +48,8 @@ class OwnershipManager(object):  # pylint:disable=too-many-instance-attributes
         self.owner_id = owner_id
         self.checkpoint_store = checkpoint_store
         self.ownership_timeout = ownership_timeout
+        self.load_balancing_strategy = load_balancing_strategy
         self.partition_id = partition_id
-        self._initializing = True
 
     async def claim_ownership(self) -> List[str]:
         """Claims ownership for this EventProcessor
@@ -121,8 +124,8 @@ class OwnershipManager(object):  # pylint:disable=too-many-instance-attributes
         ]
 
         if (
-            self._initializing
-        ):  # greedily claim all available partitions when an EventProcessor is started.
+            LoadBalancingStrategy.GREEDY is self.load_balancing_strategy
+        ):  # claim all available partitions if load balancing strategy is greedy.
             to_claim = released_partitions
             for to_claim_item in to_claim:
                 to_claim_item["owner_id"] = self.owner_id
@@ -136,12 +139,13 @@ class OwnershipManager(object):  # pylint:disable=too-many-instance-attributes
                         "owner_id": self.owner_id,
                     }
                 )
-            self._initializing = False
             if (
                 to_claim
             ):  # if no expired, released or unclaimed partitions, go ahead with balancing
                 return to_claim
 
+        # If greedy strategy is not used or doesn't find any partitions to claim
+        # load balance tries to steal at most one partition.
         released_partition_ids = [
             ownership["partition_id"] for ownership in released_partitions
         ]